Refactoring
 - Better name for some functions/variables
 - Allowing the typedLogger to have any output, its the caller's
   responsibility to restric output usage
 - Moved `typedLogger` to its own file
 - Updated tests accoringly
