+++ logp/core.go
@@ -139,7 +139,7 @@ func ConfigureWithOutputs(defaultLoggerCfg Config, outputs ...zapcore.Core) erro
 	return nil
 }
 
-// ConfigureWithTypedOutputs configures the global logger to use typed outputs.
+// ConfigureWithTypedOutput configures the global logger to use typed outputs.
 //
 // If a log entry matches the defined key/value, this entry is logged using the
 // core generated from `typedLoggerCfg`, otherwise it will be logged by all
@@ -153,15 +153,15 @@ func ConfigureWithOutputs(defaultLoggerCfg Config, outputs ...zapcore.Core) erro
 //   - `value` is the value compared against the `logKey` entry
 //   - `outputs` is a list of cores that will be added together with the core
 //     generated by `defaultLoggerCfg` as the default output for the loggger.
-func ConfigureWithTypedOutputs(defaultLoggerCfg, typedLoggerCfg Config, key, value string, outputs ...zapcore.Core) error {
+func ConfigureWithTypedOutput(defaultLoggerCfg, typedLoggerCfg Config, key, value string, outputs ...zapcore.Core) error {
 	sink, level, observedLogs, selectors, err := createSink(defaultLoggerCfg, outputs...)
 	if err != nil {
 		return err
 	}
 
-	typedCore, err := makeFileOutput(typedLoggerCfg, level)
+	typedCore, err := createLogOutput(typedLoggerCfg, level)
 	if err != nil {
-		return fmt.Errorf("could not create sensitive file output: %w", err)
+		return fmt.Errorf("could not create typed logger output: %w", err)
 	}
 
 	sink = &typedLoggerCore{
@@ -289,34 +289,6 @@ func makeEventLogOutput(cfg Config, enab zapcore.LevelEnabler) (zapcore.Core, er
 	return wrappedCore(core), nil
 }
 
-// WithFileOrStderrOutput configures the logger with an output based on cfg.
-// If neither `cfg.ToFiles` nor `cfg.ToSdterr` are true, then a noop output
-// is created. If both are true, the file output has preference.
-func WithFileOrStderrOutput(cfg Config) func(zapcore.Core) zapcore.Core {
-	var out zapcore.Core
-	var err error
-
-	switch {
-	case cfg.ToFiles:
-		out, err = makeFileOutput(cfg, cfg.Level.ZapLevel())
-	case cfg.ToStderr:
-		out, err = makeStderrOutput(cfg, cfg.Level.ZapLevel())
-	default:
-		out = zapcore.NewNopCore()
-		err = errors.New("unexpected output type, only allowed outputs are 'file' and 'stderr'")
-	}
-
-	if err != nil {
-		L().Errorf("could not create log output: %s", err)
-	}
-
-	f := func(zapcore.Core) zapcore.Core {
-		return out
-	}
-
-	return f
-}
-
 func makeFileOutput(cfg Config, enab zapcore.LevelEnabler) (zapcore.Core, error) {
 	filename := paths.Resolve(paths.Logs, filepath.Join(cfg.Files.Path, cfg.LogFilename()))
 
@@ -423,60 +395,4 @@ func (m multiCore) Sync() error {
 		}
 	}
 	return errors.Join(errs...)
-}
-
-// typedLoggerCore takes two cores and directs logs entries to one of them
-// with the value of the field defined by the pair `key` and `value`
-//
-// If `entry[key] == value` the typedCore is used, otherwise the
-// defaultCore  is used.
-// WARNING: The level of both cores must always be the same!
-// typedLoggerCore will only use the defaultCore level to decide
-// whether to log an entry or not
-type typedLoggerCore struct {
-	typedCore   zapcore.Core
-	defaultCore zapcore.Core
-	value       string
-	key         string
-}
-
-func (t *typedLoggerCore) Enabled(l zapcore.Level) bool {
-	return t.defaultCore.Enabled(l)
-}
-
-func (t *typedLoggerCore) With(fields []zapcore.Field) zapcore.Core {
-	t.defaultCore = t.defaultCore.With(fields)
-	t.typedCore = t.typedCore.With(fields)
-	return t
-}
-
-func (t *typedLoggerCore) Check(e zapcore.Entry, ce *zapcore.CheckedEntry) *zapcore.CheckedEntry {
-	if t.defaultCore.Enabled(e.Level) {
-		return ce.AddCore(e, t)
-	}
-
-	return ce
-}
-
-func (t *typedLoggerCore) Sync() error {
-	defaultErr := t.defaultCore.Sync()
-	typedErr := t.typedCore.Sync()
-
-	if defaultErr != nil || typedErr != nil {
-		return fmt.Errorf("error syncing logger. DefaultCore: '%w', typedCore: '%w'", defaultErr, typedErr)
-	}
-
-	return nil
-}
-
-func (t *typedLoggerCore) Write(e zapcore.Entry, fields []zapcore.Field) error {
-	for _, f := range fields {
-		if f.Key == t.key {
-			if f.String == t.value {
-				return t.typedCore.Write(e, fields)
-			}
-		}
-	}
-
-	return t.defaultCore.Write(e, fields)
 }
